+++ constants/consts.ts
@@ -1,5 +1,5 @@
-export const signUp = 'Dont Have An Account? Sign Up';
-export const signIn = 'already have an account? sign in';
-export const signUpText = 'Sign Up';
-export const confirmPassword = 'Confirm Password';
-export const createAcountTitle = 'CREATE ACCOUNT';
+export const SIGN_UP_MESSAGE = 'Dont Have An Account? Sign Up';
+export const SIGN_IN_MESSAGE = 'already have an account? sign in';
+export const SIGN_UP = 'Sign Up';
+export const CONFIRM_PASSWORD = 'Confirm Password';
+export const CREATE_ACCOUNT = 'CREATE ACCOUNT';
