+++ src/screens/index.ts
@@ -1,4 +1,4 @@
-export * from './Register';
+export * from './SignUp';
 export * from './DivingMap';
-export * from './LoginScreen';
-export * from './AppendDive';
+export * from './Login';
+export * from './AddDive';
